src/pyproject.toml
[tool.poetry]
name = "e-noewshop-users"
version = "0.0.1"
description = ""
authors = ["gustavoDanielT <gustavo.toledo@shipay.com.br>"]

[tool.poetry.dependencies]
python = "^3.7"
SQLAlchemy = "^1.4.39"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
dependency-injector = "^4.40.0"
aiokafka = "^0.7.2"
pycpfcnpj = "^1.7.0"

e-nowshop-models = {path = "/Users/gustavotoledo/Project/e-nowshop/e-nowshop-models"}

httpx = "^0.23.0"
asyncpg = "^0.26.0"
jwt = "^1.3.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
